restructured addMamacMeter.js and extracted meter information in backend
